Move pdf folder into /common
