+++ src/common/website-helper/pdf/pdf-parser/horaire/HoraireCours.ts
@@ -1,4 +1,4 @@
-import { CourseCodeValidationPipe } from '../../../common/pipes/models/course/course-code-validation-pipe';
+import { CourseCodeValidationPipe } from '../../../../pipes/models/course/course-code-validation-pipe';
 import { Group, IGroup } from './Group';
 import { IHoraireCours } from './horaire-cours.types';
 import { Period } from './Period';
